changed the code so now delete is within the edit page

+++ src/main/java/com/example/wagster/models/Location.java
@@ -31,7 +31,7 @@ public class Location {
     private String description;
 
     @Column(nullable = false)
-    private short rating;
+    private float rating;
 
     @ManyToOne
     @JoinColumn( name = "user_id")
